Added Lazy Librarian
Tested on Linux Mint & Raspbian

Changes made to allow for systemD startup for this and future applications.

Defaulted to be accessible from other devices on the network but can be changed to be only accessible from locahost.

+++ inc/app-init-remove.sh
@@ -2,9 +2,13 @@
 echo
 sleep 1
 
-echo -e $YELLOW'--->Removing '$APPTITLE' Autostart scripts...'$ENDCOLOR
-sudo update-rc.d -f $APPNAME remove || { echo -e $RED'Warning! update-rc.d remove failed.'$ENDCOLOR ; }
-sudo rm /etc/init.d/$APPNAME || { echo -e $RED'Warning! Removing init script failed.'$ENDCOLOR ; }
+echo -e $YELLOW'--->Removing '$APPTITLE' Init.d Autostart scripts...'$ENDCOLOR
+
+if [ -f /etc/init.d/$APPNAME ]; then
+	sudo update-rc.d -f $APPNAME remove || { echo -e $RED'Warning! update-rc.d remove failed.'$ENDCOLOR ; }
+	sudo rm /etc/init.d/$APPNAME || { echo -e $RED'Warning! Removing init script failed.'$ENDCOLOR ; }
+fi
+
 if [ -f /etc/default/$APPNAME ]; then
 	sudo rm /etc/default/$APPNAME || { echo -e $RED'Warning! Removing default script failed.'$ENDCOLOR ; }
 fi
